Move newly created css files to keep folder clarity and readability. #682
